Simplified and moved Tree printer Test
